+++ src/game/systems/morphogenesis-system.cpp
@@ -25,6 +25,6 @@ morphogenesis_system::morphogenesis_system(entity::registry& registry):
 	updatable_system(registry)
 {}
 
-void morphogenesis_system::update(double t, double dt)
+void morphogenesis_system::update(float t, float dt)
 {}
